refactor(fcm): Move device functions to helper
